requirements.txt
altgraph==0.17.3
bottle==0.12.25
cffi==1.15.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pefile==2023.2.7
proxy-tools==0.1.0
pycparser==2.21
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pythonnet==3.0.1
pywebview==4.0.2
pywin32-ctypes==0.2.0
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
